requirements.txt
streamlit>=1.36
anthropic>=0.36
trafilatura>=1.9.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
python-dotenv>=1.0.1
pydub>=0.25.1
requests>=2.32
tqdm>=4.66
elevenlabs>=1.1.0
